Added tiedotteet as an app to the project
This required consolidating settings.py, including namespaced tiedotteet urls to urls.py
and rerouting some urls in the tiedotteet app templates and webpack.config.js. Namely renamed
/public folder to 'static' and appended tiedotteet: to each {% url %} template tag. TODO:
common auth and adding matrikkeli as an app to the project. Another way to do this would
have been to use multiple databases and designate one auth_db. The drawback with having one
big project is that the github repos probably need to be consolidated aswell. NOTE: to use
the tiedotteet app nodejs and yarn are needed. cd to /client and do 'yarn install' and then 'yarn build'